BLAST/requirements.txt
biopython==1.81
pandas==1.4.4
numpy==1.26.4